#27 - moves ui test to misc module
